requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==24.2.0
cachetools==5.3.1
cassandra-driver==3.28.0
cassio==0.1.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.5.14
distro==1.9.0
exceptiongroup==1.1.3
fastapi==0.109.2
frozenlist==1.4.0
geomet==0.2.1.post1
google-auth==2.22.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.13.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==27.2.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.8
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.24.4
oauthlib==3.2.2
openai==1.12.0
orjson==3.9.15
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.2.1
pydantic-core==2.6.1
pydantic-settings==2.0.3
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.10.0
typing-inspect==0.9.0
urllib3==1.26.18
uvicorn==0.23.2
websocket-client==1.6.2
yarl==1.9.2
